refactor(tests): criar tests/unit/conftest.py e mover mock_sqlalchemy_engine
Move mock_sqlalchemy_engine do conftest raiz para tests/unit/conftest.py
(escopo mais adequado) e adiciona mock_psycopg2_conn como fixture
compartilhada para eliminar boilerplate duplicado nos próximos commits.

Corrige também test_connection_string_with_cloud_sql_proxy que falhava
quando executado após testes de integração (DATABASE_URL vazava do
conftest de integração por ser session-scoped).

Refs: #136
