+++ tests/unit/conftest.py
@@ -1,0 +1,26 @@
+"""Shared fixtures for unit tests."""
+
+from unittest.mock import MagicMock
+
+import pytest
+
+
+@pytest.fixture
+def mock_sqlalchemy_engine():
+    """Build a mock SQLAlchemy engine that supports `with engine.begin() as conn`."""
+    mock_conn = MagicMock()
+    mock_engine = MagicMock()
+    mock_engine.begin.return_value.__enter__ = MagicMock(return_value=mock_conn)
+    mock_engine.begin.return_value.__exit__ = MagicMock(return_value=False)
+    return mock_engine, mock_conn
+
+
+@pytest.fixture
+def mock_psycopg2_conn():
+    """Reusable mock psycopg2 connection + cursor with context manager support."""
+    mock_cursor = MagicMock()
+    mock_cursor.__enter__ = MagicMock(return_value=mock_cursor)
+    mock_cursor.__exit__ = MagicMock(return_value=False)
+    mock_conn = MagicMock()
+    mock_conn.cursor.return_value = mock_cursor
+    return mock_conn, mock_cursor
